+++ src/_pytask/nodes.py
@@ -8,7 +8,6 @@ from pathlib import Path
 from typing import Any
 from typing import Callable
 from typing import Dict
-from typing import Iterable
 from typing import List
 from typing import Optional
 from typing import Tuple
@@ -20,38 +19,6 @@ from _pytask.exceptions import NodeNotFoundError
 
 if TYPE_CHECKING:
     from _pytask.mark import Mark
-
-
-def depends_on(
-    objects: Any | Iterable[Any] | dict[Any, Any]
-) -> Any | Iterable[Any] | dict[Any, Any]:
-    """Specify dependencies for a task.
-
-    Parameters
-    ----------
-    objects : Union[Any, Iterable[Any], Dict[Any, Any]]
-        Can be any valid Python object or an iterable of any Python objects. To be
-        valid, it must be parsed by some hook implementation for the
-        :func:`pytask.hookspecs.pytask_collect_node` entry-point.
-
-    """
-    return objects
-
-
-def produces(
-    objects: Any | Iterable[Any] | dict[Any, Any]
-) -> Any | Iterable[Any] | dict[Any, Any]:
-    """Specify products of a task.
-
-    Parameters
-    ----------
-    objects : Union[Any, Iterable[Any], Dict[Any, Any]]
-        Can be any valid Python object or an iterable of any Python objects. To be
-        valid, it must be parsed by some hook implementation for the
-        :func:`pytask.hookspecs.pytask_collect_node` entry-point.
-
-    """
-    return objects
 
 
 class MetaNode(metaclass=ABCMeta):
@@ -119,10 +86,8 @@ class FilePathNode(MetaNode):
 
     name = attr.ib(type=str)
     """str: Name of the node which makes it identifiable in the DAG."""
-
     value = attr.ib(type=Path)
     """Any: Value passed to the decorator which can be requested inside the function."""
-
     path = attr.ib(type=Path)
     """pathlib.Path: Path to the FilePathNode."""
 
@@ -144,25 +109,3 @@ class FilePathNode(MetaNode):
             raise NodeNotFoundError
         else:
             return str(self.path.stat().st_mtime)
-
-
-def find_duplicates(x: Iterable[Any]) -> set[Any]:
-    """Find duplicated entries in iterable.
-
-    Examples
-    --------
-    >>> find_duplicates(["a", "b", "a"])
-    {'a'}
-    >>> find_duplicates(["a", "b"])
-    set()
-
-    """
-    seen = set()
-    duplicates = set()
-
-    for i in x:
-        if i in seen:
-            duplicates.add(i)
-        seen.add(i)
-
-    return duplicates